// src/app/config/types/app-config.type.ts
export type AppConfig = {
  fallbackLanguage: string;
  headerLanguage: string;
  frontendUrl: string;
  apiUrl: string;
  apiPort: number;
  name: string;
  dbPoolMax: number;
  dbName: string;
  dbPassword: string;
  dbUsername: string;
  dbHost: string;
  filesDir: string;
  firebaseProjectId: string;
  firebaseClientEmail: string;
  firebasePrivateKey: string;
};